angular UI added

+++ VidySphere.Domain/Entities/Course.cs
@@ -12,5 +12,5 @@ public class Course : BaseEntity
     public string Title { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
 
-    public int InstructorId { get; set; }
+    public long InstructorId { get; set; }
 }
